pyproject.toml
[tool.poetry]
name = "ac_ai_api"
version = "0.1.0"
description = "first poetry project"
authors = ["Ailenei Dan <dan.ailenei@rover.com>"]

[tool.poetry.dependencies]
python = "3.8"
django = "*"
django-extensions = "*"
hunter = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
djangorestframework = "^3.11.0"
ipython = "^7.11.1"
django-rest-knox = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
